separated stackable inspector options
